Split up the tests for ApplyCampaign into separate file
